# docs/openapi-v1.json
{
  "components": {
    "schemas": {
...
      "BenchmarkOut": {
        "additionalProperties": false,
        "description": "`GET /benchmarks` \u2014 cohort baselines.",
        "properties": {
          "cohort": {
            "title": "Cohort",
            "type": "string"
          },
          "fell_back_to_covariate_model": {
            "default": false,
            "title": "Fell Back To Covariate Model",
            "type": "boolean"
          },
          "method": {
            "title": "Method",
            "type": "string"
          },
          "metric": {
            "title": "Metric",
            "type": "string"
          },
          "n_peers": {
            "minimum": 0.0,
            "title": "N Peers",
            "type": "integer"
          },
          "peer_mad": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Peer Mad"
          },
          "peer_median": {
            "title": "Peer Median",
            "type": "number"
          }
        },
        "required": [
          "cohort",
          "metric",
          "n_peers",
          "peer_median",
          "method"
        ],
        "title": "BenchmarkOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PolicyActivateIn": {
        "additionalProperties": false,
        "properties": {
          "effective_from": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective From"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "title": "PolicyActivateIn",
        "type": "object"
      },
      "PolicyActivateOut": {
        "additionalProperties": false,
        "properties": {
          "activated": {
            "title": "Activated",
            "type": "boolean"
          },
          "ledger_entry_hash": {
            "title": "Ledger Entry Hash",
            "type": "string"
          },
          "previous_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous Id"
          },
          "profile_id": {
            "title": "Profile Id",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "profile_id",
          "version",
          "activated",
          "ledger_entry_hash"
        ],
        "title": "PolicyActivateOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Supervisory Analytics Tool for SOC Assessment (SAT-SA) for NCIIPC.\n\nOutputs are **prioritised hypotheses with evidence**, never verdicts or\ncompliance grades. `Not assessable` is a first-class outcome, distinct from\n\"low risk\".\n\nFully offline: no cloud APIs, no CDN, no generative AI in the decision path.\n\n## Reading this contract\n\n`Not assessable` is a **value**, not an absence. Every score field is nullable\nand `null` means \"could not be assessed\" \u2014 never \"scored zero\". The frontend must\nnot coalesce `null` to `0`; a dimension with no data and a dimension with a\nmeasured score of zero are different claims.\n\n`POST /verdicts` records **a supervisor's own conclusion**, written by a human.\nNo Orion-derived payload in this API contains a verdict, a compliance grade, or a\npass/fail.\n\nRoutes marked `503` are part of this frozen contract but their backing service is\nnot written yet. The response body names the service and the phase that delivers\nit. They are listed rather than omitted so the surface does not change shape\nmid-project; `GET /api/v1` splits them into `live` and `pending`.",
    "title": "Orion (SAT-SA) Core Backend",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/benchmarks": {
      "get": {
        "description": "Peer distributions, never thresholds. A median and a MAD describe a population; they do not say what a population *should* look like, and nothing here is a target. With no period given, the most recent scoring run is used.",
        "operationId": "benchmarks_api_v1_benchmarks_get",
        "parameters": [
          {
            "description": "egi, nsi, dts or sap",
            "in": "query",
            "name": "metric",
            "required": false,
            "schema": {
              "default": "egi",
              "description": "egi, nsi, dts or sap",
              "title": "Metric",
              "type": "string"
            }
          },
          {
            "description": "Entity to exclude, so it is not compared against itself",
            "in": "query",
            "name": "exclude",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Entity to exclude, so it is not compared against itself",
              "title": "Exclude"
            }
          },
          {
            "description": "Inclusive start, YYYY-MM-DD",
            "in": "query",
            "name": "period_start",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Inclusive start, YYYY-MM-DD",
              "title": "Period Start"
            }
          },
          {
            "description": "Inclusive end, YYYY-MM-DD",
            "in": "query",
            "name": "period_end",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Inclusive end, YYYY-MM-DD",
              "title": "Period End"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BenchmarkOut"
                  },
                  "title": "Response Benchmarks Api V1 Benchmarks Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cohort baselines for the executive view",
        "tags": [
          "policy"
        ]
      }
    },
...
    "/api/v1/policy-profiles/{profile_id}/activate": {
      "post": {
        "description": "Does not rewrite history. Scores already stored keep the policy hash they were computed under, so a stored figure is always interpretable against the profile that produced it. What changes is what the *next* run means, which is why `previous_id` comes back in the response.",
        "operationId": "activate_profile_api_v1_policy_profiles__profile_id__activate_post",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "title": "Profile Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Actor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Actor"
            }
          },
          {
            "in": "header",
            "name": "X-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Role"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PolicyActivateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyActivateOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "description": "Role may not activate a policy profile"
          },
          "404": {
            "description": "No such profile"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Make a profile the one new runs are computed under. Ledgered.",
        "tags": [
          "policy"
        ]
      }
    },
...
  },
...
}